refactor: move static helpers out to GLPreprocessorHelpers, decouple them from state
