+++ src/main/java/titanicsend/pattern/glengine/GLPreprocessor.java
@@ -1,15 +1,16 @@
 package titanicsend.pattern.glengine;
 
+import static titanicsend.pattern.glengine.GLPreprocessorHelpers.*;
+
 import heronarts.lx.parameter.BooleanParameter;
-import heronarts.lx.parameter.BoundedParameter;
 import heronarts.lx.parameter.CompoundParameter;
 import heronarts.lx.parameter.LXParameter;
 import java.io.*;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import titanicsend.pattern.jon.TEControlTag;
 import titanicsend.pattern.yoffa.shader_engine.ShaderUtils;
 import titanicsend.pattern.yoffa.shader_engine.Uniforms;
 
@@ -21,6 +22,68 @@ public class GLPreprocessor {
   private boolean foundInclude = false;
   private int lineCount = 0;
   private boolean isDriftModeShader = false;
+
+  /**
+   * Preprocess the shader, expanding #includes and handling our TE-specific control and texture
+   * configuration #pragmas. Note that this preprocessor doesn't support the old method of adding
+   * extra controls. TODO - should we add this as an option?
+   */
+  public String preprocessShader(File shaderFile, List<ShaderConfiguration> parameters)
+      throws Exception {
+    String shaderBody = ShaderUtils.loadResource(shaderFile);
+    return preprocessShader(shaderBody, parameters);
+  }
+
+  public String preprocessShader(String shaderBody, List<ShaderConfiguration> parameters)
+      throws Exception {
+    lineCount = 0;
+    try {
+      int depth = 0;
+      while (true) {
+        if (depth >= MAX_INCLUDE_DEPTH) {
+          throw new RuntimeException("Exceeded maximum #include depth of " + MAX_INCLUDE_DEPTH);
+        }
+        shaderBody = expandIncludes(shaderBody);
+        depth++;
+        if (!this.foundInclude) {
+          break;
+        }
+      }
+      shaderBody = legacyPreprocessor(shaderBody, parameters);
+
+      // VSCode Shadertoy extension support - parse '#iUniform' syntax as an alternate
+      // to '#pragma TEControl.NAME RANGE' syntax.
+      List<ShaderConfiguration> configsFromIUniforms = parseIUniforms(shaderBody);
+      parameters.addAll(configsFromIUniforms);
+      // After extracting those, remove all lines beginning with #iUniform.
+      shaderBody = removeIUniformLines(shaderBody);
+
+      // Extract configs specified using #pragma
+      List<ShaderConfiguration> configsFromPragmas = parsePragmas(shaderBody);
+      for (ShaderConfiguration control : configsFromPragmas) {
+        // If there was a pragma indicating DRIFT, we need to maintain some state so
+        // we can pass it to the template shader.
+        if (control.opcode == ShaderConfigOpcode.SET_TRANSLATE_MODE_DRIFT) {
+          isDriftModeShader = true;
+          break;
+        }
+      }
+      parameters.addAll(configsFromPragmas);
+    } catch (Exception e) {
+      throw new Exception("Shader Preprocessor Error. " + e.getMessage());
+    }
+
+    // in drift mode shaders, x/y translate controls set movement direction and speed rather than
+    // absolute offset.  Define a constant to tell the shader framework what we want.
+    if (isDriftModeShader) {
+      shaderBody = "#define TE_NOTRANSLATE\n" + shaderBody;
+    }
+
+    // combine the fragment shader code with the framework template
+    shaderBody =
+        getFragmentShaderTemplate().replace(ShaderUtils.SHADER_BODY_PLACEHOLDER, shaderBody);
+    return shaderBody;
+  }
 
   // Expand #include statements in the shader code.  Handles nested includes
   // up to MAX_INCLUDE_DEPTH (defaults to 10 levels.)
@@ -34,7 +97,7 @@ public class GLPreprocessor {
       if (line.startsWith("#include")) {
         foundInclude = true;
         try {
-          String filename = getFileName(line.substring("#include ".length(), line.length()));
+          String filename = getFileName(line.substring("#include ".length()));
 
           BufferedReader fileReader = new BufferedReader(new FileReader(filename));
           String fileLine;
@@ -59,139 +122,11 @@ public class GLPreprocessor {
     return output.toString();
   }
 
-  // Converts strings from control definition #pragmas to shader configuration
-  // opcode values
-  public static ShaderConfigOpcode opcodeFromString(String str) {
-    return switch (str) {
-      case "auto" -> ShaderConfigOpcode.AUTO;
-      case "Value" -> ShaderConfigOpcode.SET_VALUE;
-      case "Range" -> ShaderConfigOpcode.SET_RANGE;
-      case "Label" -> ShaderConfigOpcode.SET_LABEL;
-      case "Exponent" -> ShaderConfigOpcode.SET_EXPONENT;
-      case "NormalizationCurve" -> ShaderConfigOpcode.SET_NORMALIZATION_CURVE;
-      case "Disable" -> ShaderConfigOpcode.DISABLE;
-      case "NORMAL" -> ShaderConfigOpcode.SET_TRANSLATE_MODE_NORMAL;
-      case "DRIFT" -> ShaderConfigOpcode.SET_TRANSLATE_MODE_DRIFT;
-      default -> throw new IllegalArgumentException("Unknown configuration operation: " + str);
-    };
-  }
-
-  // Parse a control definition #pragma and add it to the list of shader
-  // configuration parameters
-  public void parseControl(String[] line, List<ShaderConfiguration> parameters) {
-    ShaderConfiguration control = new ShaderConfiguration();
-    // tokenize the first element, dividing by periods
-    String[] tokens = line[0].split("\\.");
-
-    // token[0] will always be "TEControl" to specify common control configuration
-    //
-    // token[1] will be the control name, which should correspond to a TEControlTag entry
-    // or 'TranslateMode', which specifies how the pattern should handle x/y translation.
-    if (tokens[1].equals("TranslateMode")) {
-      // NOTE: This affects pattern subclassing, so must be handled separately
-      // from (and prior to) "normal" control configurations during initialization.
-      control.opcode = opcodeFromString(line[1]);
-      isDriftModeShader = (control.opcode == ShaderConfigOpcode.SET_TRANSLATE_MODE_DRIFT);
-    } else {
-      control.parameterId = TEControlTag.valueOf(tokens[1].toUpperCase());
-      control.name = control.parameterId.getLabel();
-      // the third token is the operation to be performed on the specified control
-
-      control.opcode = opcodeFromString(tokens[2]);
-
-      // now we need to parse the rest of the tokens, which will depend on the operation
-      switch (control.opcode) {
-        case SET_VALUE, SET_EXPONENT -> control.value = Double.parseDouble(line[1]);
-        case SET_RANGE -> {
-          // tokenize line[1] by commas to get setRange() parameters
-          String[] range = line[1].split(",");
-          if (range.length != 3) {
-            throw new IllegalArgumentException("Invalid range specification: " + line[1]);
-          }
-          control.value = Double.parseDouble(range[0]);
-          control.v1 = Double.parseDouble(range[1]);
-          control.v2 = Double.parseDouble(range[2]);
-        }
-        case SET_LABEL -> control.name = stringCleanup(line[1]);
-        case SET_NORMALIZATION_CURVE -> control.normalizationCurve =
-            BoundedParameter.NormalizationCurve.valueOf(line[1].toUpperCase());
-      }
-    }
-    parameters.add(control);
-  }
-
-  // Parse a texture definition #pragma and add it to the list of shader
-  // configuration parameters
-  public void parseTextures(String[] line, List<ShaderConfiguration> parameters) {
-    ShaderConfiguration control = new ShaderConfiguration();
-    control.opcode = ShaderConfigOpcode.SET_TEXTURE;
-
-    // the last character of token 0 is the integer channel identifier
-    // valid channels are 1-9.  Channel 0 is reserved for audio input.
-    control.textureChannel = Integer.parseInt(line[0].substring(line[0].length() - 1));
-    if (control.textureChannel == 0) {
-      throw new IllegalArgumentException(
-          "iChannel0 is reserved for system audio. Use channels 1-9 for textures.");
-    }
-
-    // token 1 is the texture file name.
-    control.name = getFileName(line[1]);
-    parameters.add(control);
-  }
-
-  public void parseClassName(String[] line, List<ShaderConfiguration> parameters) {
-    ShaderConfiguration control = new ShaderConfiguration();
-    control.opcode = ShaderConfigOpcode.SET_CLASS_NAME;
-
-    // token 1 is the desired class name and default pattern name
-    control.name = stringCleanup(line[1]);
-    parameters.add(control);
-  }
-
-  public void parseLXCategory(String[] line, List<ShaderConfiguration> parameters) {
-    ShaderConfiguration control = new ShaderConfiguration();
-    control.opcode = ShaderConfigOpcode.SET_LX_CATEGORY;
-
-    // since spaces are permissible in category names, we need to merge line's elements
-    // from index 1 to the end back into a single string
-    control.name = stringCleanup(String.join(" ", Arrays.copyOfRange(line, 1, line.length)));
-    parameters.add(control);
-  }
-
-  private static String stringCleanup(String str) {
-    // clean up delimiters
-    if (str.startsWith("\"") && str.endsWith("\"")) {
-      str = str.substring(1, str.length() - 1);
-    }
-    return str.trim();
-  }
-
-  // Convert an input token to a valid filename, removing any delimiters and
-  // checking to see that the file actually exists.
-  private static String getFileName(String str) {
-    str = stringCleanup(str);
-
-    // if name is enclosed in angle brackets, prefix with default resource path
-    // to save repetitive typing
-    if (str.startsWith("<") && str.endsWith(">")) {
-      str = str.substring(1, str.length() - 1);
-      str = ShaderUtils.SHADER_PATH + str;
-      // cleanup again in case there were spaces or more quotes
-      str = stringCleanup(str);
-    }
-
-    // check to see if the file actually exists
-    File f = new File(str);
-    if (!f.exists()) {
-      throw new IllegalArgumentException("File " + str + " not found.");
-    }
-    return str;
-  }
-
   // Parse #pragma statements in the shader code, subdividing them into shader
   // configuration parameters and texture definitions.  Note that GLSL requires
   // us to ignore any #pragma statements that we don't recognize.
-  public void parsePragmas(String input, List<ShaderConfiguration> parameters) {
+  public List<ShaderConfiguration> parsePragmas(String input) {
+    List<ShaderConfiguration> parameters = new ArrayList<>();
     Pattern pattern = Pattern.compile("^\\s*#pragma.*", Pattern.MULTILINE);
     Matcher matcher = pattern.matcher(input);
 
@@ -205,25 +140,25 @@ public class GLPreprocessor {
         // discard the #pragma token
         tokens = Arrays.copyOfRange(tokens, 1, tokens.length);
 
-        // Common controls configuration
         String pragma = tokens[0].toLowerCase();
         if (pragma.startsWith("tecontrol.")) {
-          parseControl(tokens, parameters);
-        }
-        // Texture channel definition
-        else if (pragma.startsWith("ichannel")) {
-          parseTextures(tokens, parameters);
-        }
-        // name of class/pattern in UI
-        else if (pragma.equals("name")) {
-          parseClassName(tokens, parameters);
-        }
-        // set LXCategory for pattern
-        else if (pragma.equals("lxcategory")) {
-          parseLXCategory(tokens, parameters);
-        }
-        // auto keyword forces use of automatic class generation system
-        else if (pragma.equals("auto")) {
+          // Common controls configuration
+          ShaderConfiguration control = parseControl(tokens);
+          parameters.add(control);
+        } else if (pragma.startsWith("ichannel")) {
+          // Texture channel definition
+          ShaderConfiguration control = parseTextures(tokens);
+          parameters.add(control);
+        } else if (pragma.equals("name")) {
+          // name of class/pattern in UI
+          ShaderConfiguration control = parseClassName(tokens);
+          parameters.add(control);
+        } else if (pragma.equals("lxcategory")) {
+          // set LXCategory for pattern
+          ShaderConfiguration control = parseLXCategory(tokens);
+          parameters.add(control);
+        } else if (pragma.equals("auto")) {
+          // auto keyword forces use of automatic class generation system
           ShaderConfiguration p = new ShaderConfiguration();
           p.opcode = ShaderConfigOpcode.AUTO;
           parameters.add(p);
@@ -232,153 +167,7 @@ public class GLPreprocessor {
         throw new RuntimeException("Error in " + matcher.group() + "\n" + e.getMessage());
       }
     }
-  }
-
-  // Parse #iUniform directives (to be compatible with VSCode ShaderToy extension).
-  public void parseIUniforms(String input, List<ShaderConfiguration> parameters) {
-    Pattern pattern = Pattern.compile("^\\s*#iUniform.*", Pattern.MULTILINE);
-    Matcher matcher = pattern.matcher(input);
-
-    while (matcher.find()) {
-      try {
-        // tokenize the line, dividing first by whitespace and parentheses
-        // NOTE: trim leading/trailing whitespace, so that leading/trailing
-        // newline matches don't screw up our parsing.
-        String[] parts = matcher.group().trim().split("=");
-        if (parts.length != 2) {
-          throw new Exception("Expected 2 parts delimited by '=', but found: " + parts.length);
-        }
-        String[] lhsTokens = parts[0].split("\\s|\\(|\\)");
-        String rhs = parts[1];
-
-        //        System.out.println("(in) LHS: " + Arrays.toString(lhsTokens));
-        if (lhsTokens.length != 3) {
-          throw new Exception(
-              "Expected 3 LHS tokens delimited by whitespace, but found: "
-                  + lhsTokens.length
-                  + " in string'"
-                  + parts[0]
-                  + "'");
-        } else if (!lhsTokens[0].equals("#iUniform")) {
-          throw new Exception("Expected first token on LHS to be '#iUniform'");
-        }
-        String varType = lhsTokens[1];
-        String varName = lhsTokens[2];
-        //        System.out.println("(out) LHS: '"+varType+" "+varName+"'");
-
-        if (varType.equals("float")) {
-          /*
-           * Regular expression pattern to extract three floating-point values from strings in the format:
-           * "<value> in {<lowerBound>,<upperBound>}"
-           *
-           * Pattern: (\\d*\\.\\d*|\\d+\\.?)\\s*in\\s*\\{\\s*(\\d*\\.\\d*|\\d+\\.?)\\s*,\\s*(\\d*\\.\\d*|\\d+\\.?)\\s*\\}
-           *
-           * Breakdown:
-           * 1. (\\d*\\.\\d*|\\d+\\.?) - Captures a floating-point number in GLSL notation:
-           *    - \\d*\\.\\d* matches numbers like ".5", "0.5", or "1.0"
-           *    - \\d+\\.? matches numbers like "1" or "1."
-           *    - The | (OR) operator allows either format
-           *
-           * 2. \\s*in\\s* - Matches the word "in" with optional whitespace before and after
-           *
-           * 3. \\{\\s* - Matches the opening curly brace with optional whitespace after
-           *
-           * 4. The same floating-point pattern is repeated for the lower bound
-           *
-           * 5. \\s*,\\s* - Matches the comma separator with optional whitespace
-           *
-           * 6. The floating-point pattern is repeated again for the upper bound
-           *
-           * 7. \\s*\\} - Matches the closing curly brace with optional whitespace before
-           *
-           * Capturing groups:
-           * - Group 1: The initial value
-           * - Group 2: The lower bound
-           * - Group 3: The upper bound
-           */
-          Pattern floatRangePattern =
-              Pattern.compile(
-                  "(-?\\d*\\.\\d*|\\d+\\.?)\\s*"
-                      + "in\\s*\\{\\s*(-?\\d*\\.\\d*|\\d+\\.?)"
-                      + "\\s*,\\s*(-?\\d*\\.\\d*|\\d+\\.?)"
-                      + "\\s*}" // Removed the \n and escaped the closing brace properly
-                  );
-
-          Matcher floatRangeMatcher = floatRangePattern.matcher(rhs);
-
-          if (!floatRangeMatcher.find()) {
-            throw new RuntimeException("float range didn't match: [" + rhs + "]");
-          }
-
-          float rangeDefault = parseGlslFloat(floatRangeMatcher.group(1));
-          float rangeLower = parseGlslFloat(floatRangeMatcher.group(2));
-          float rangeUpper = parseGlslFloat(floatRangeMatcher.group(3));
-
-          ShaderConfiguration control = new ShaderConfiguration();
-
-          String tagName = varName.trim().toUpperCase();
-          if (tagName.startsWith("I")) {
-            // "iSpeed" should look up tag "SPEED"
-            tagName = tagName.substring(1);
-          }
-          try {
-
-            control.opcode = ShaderConfigOpcode.SET_RANGE;
-            control.parameterId = TEControlTag.valueOf(tagName);
-            control.name = control.parameterId.getLabel();
-            control.value = rangeDefault;
-            control.v1 = rangeLower;
-            control.v2 = rangeUpper;
-            parameters.add(control);
-          } catch (IllegalArgumentException exception) {
-            System.out.println("Unsupported tag name: " + varName);
-          }
-          // System.out.println("(out) RHS: " + rangeDefault + ", " + rangeLower + ", " +
-          // rangeUpper);
-        } else if (varType.equals("vec2") || varType.equals("vec3")) {
-          Pattern vecPattern = Pattern.compile("(vec\\d+)\\(([^)]*)\\)");
-          Matcher vecMatcher = vecPattern.matcher(rhs);
-
-          if (!vecMatcher.find()) {
-            throw new RuntimeException("vec matcher failed: " + rhs);
-          }
-          // String dataType = vecMatcher.group(1); // "vec3" or "vec2"
-          String paramsContent = vecMatcher.group(2); // ".226,.046,.636" or ".1, 2.0, -5"
-          // If you need to further process the parameters
-          String[] params = paramsContent.split(",");
-          Float[] values =
-              Arrays.stream(params).map(GLPreprocessor::parseGlslFloat).toArray(Float[]::new);
-
-          System.out.println("(out) RHS: " + Arrays.toString(values));
-
-          // TODO(look): do I need to update any TEControls? I think I only use vec2/vec3 to
-          // replicate color/translate.
-        } else if (varType.equals("color3")) {
-          // no-op
-        } else {
-          throw new RuntimeException("iUniform data type not yet inmplemented: " + varType);
-        }
-      } catch (Exception e) {
-        throw new RuntimeException("Error in " + matcher.group() + "\n" + e.getMessage());
-      }
-    }
-  }
-
-  private static float parseGlslFloat(String s) {
-    if (s.startsWith(".")) {
-      return Float.parseFloat("0" + s);
-    } else if (s.endsWith(".")) {
-      return Float.parseFloat(s + "0");
-    }
-    return Float.parseFloat(s);
-  }
-
-  private static String removeIUniformLines(String fileContent) {
-    String[] lines =
-        Arrays.stream(fileContent.split("\\n"))
-            .filter(line -> !line.contains("#iUniform"))
-            .toArray(String[]::new);
-    return String.join("\n", lines);
+    return parameters;
   }
 
   public static String getFragmentShaderTemplate() {
@@ -436,54 +225,5 @@ public class GLPreprocessor {
     finalShader.append(shaderCode);
 
     return finalShader.toString();
-  }
-
-  /**
-   * Preprocess the shader, expanding #includes and handling our TE-specific control and texture
-   * configuration #pragmas. Note that this preprocessor doesn't support the old method of adding
-   * extra controls. TODO - should we add this as an option?
-   */
-  public String preprocessShader(File shaderFile, List<ShaderConfiguration> parameters)
-      throws Exception {
-    String shaderBody = ShaderUtils.loadResource(shaderFile);
-    return preprocessShader(shaderBody, parameters);
-  }
-
-  public String preprocessShader(String shaderBody, List<ShaderConfiguration> parameters)
-      throws Exception {
-    lineCount = 0;
-    try {
-      int depth = 0;
-      while (true) {
-        if (depth >= MAX_INCLUDE_DEPTH) {
-          throw new RuntimeException("Exceeded maximum #include depth of " + MAX_INCLUDE_DEPTH);
-        }
-        shaderBody = expandIncludes(shaderBody);
-        depth++;
-        if (!this.foundInclude) {
-          break;
-        }
-      }
-      shaderBody = legacyPreprocessor(shaderBody, parameters);
-
-      // VSCode Shadertoy extension support
-      parseIUniforms(shaderBody, parameters);
-      shaderBody = removeIUniformLines(shaderBody);
-
-      parsePragmas(shaderBody, parameters);
-    } catch (Exception e) {
-      throw new Exception("Shader Preprocessor Error. " + e.getMessage());
-    }
-
-    // in drift mode shaders, x/y translate controls set movement direction and speed rather than
-    // absolute offset.  Define a constant to tell the shader framework what we want.
-    if (isDriftModeShader) {
-      shaderBody = "#define TE_NOTRANSLATE\n" + shaderBody;
-    }
-
-    // combine the fragment shader code with the framework template
-    shaderBody =
-        getFragmentShaderTemplate().replace(ShaderUtils.SHADER_BODY_PLACEHOLDER, shaderBody);
-    return shaderBody;
   }
 }
